# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "ODBUS Consumer APIs",
        "description": "L5 Swagger OpenApi description for ODBUS Consumer APIs",
        "version": "0.1"
    },
    "paths": {
...
        "/api/allReviews": {
            "get": {
                "tags": [
                    "All Reviews"
                ],
                "summary": "get all Reviews of a customer",
                "description": "get all Reviews of a customer",
                "operationId": "264ef0c3ab1ded9373510646f6b003d1",
                "responses": {
                    "200": {
                        "description": "Get all Reviews of a customer"
                    },
                    "206": {
                        "description": "validation error"
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "401": {
                        "description": "Unauthorized access"
                    },
                    "404": {
                        "description": "No record found"
                    },
                    "500": {
                        "description": "Internal server error"
                    },
                    "502": {
                        "description": "Bad gateway"
                    },
                    "503": {
                        "description": "Service unavailable"
                    },
                    "504": {
                        "description": "Gateway timeout"
                    }
                },
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            }
        },
        "/api/AddReview": {
            "post": {
                "tags": [
                    "Add Review by a customer"
                ],
                "summary": "Add Review of a customer",
                "description": "Add Review of a customer",
                "operationId": "6fe632e412681aca0c60ee00597f524a",
                "parameters": [
                    {
                        "name": "pnr",
                        "in": "query",
                        "description": "pnr number",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": "21212121"
                        }
                    },
                    {
                        "name": "bus_id",
                        "in": "query",
                        "description": "bus Id",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": "7"
                        }
                    },
                    {
                        "name": "users_id",
                        "in": "query",
                        "description": "users id",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": "2"
                        }
                    },
                    {
                        "name": "reference_key",
                        "in": "query",
                        "description": "reference key",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "abcd@gmail.com"
                        }
                    },
                    {
                        "name": "rating_overall",
                        "in": "query",
                        "description": "rating overall",
                        "required": true,
                        "schema": {
                            "type": "number",
                            "example": "4"
                        }
                    },
                    {
                        "name": "rating_comfort",
                        "in": "query",
                        "description": "rating comfort",
                        "required": true,
                        "schema": {
                            "type": "number",
                            "example": "4"
                        }
                    },
                    {
                        "name": "rating_clean",
                        "in": "query",
                        "description": "rating_clean",
                        "required": true,
                        "schema": {
                            "type": "number",
                            "example": "4"
                        }
                    },
                    {
                        "name": "rating_behavior",
                        "in": "query",
                        "description": "rating_behavior",
                        "required": true,
                        "schema": {
                            "type": "number",
                            "example": "5"
                        }
                    },
                    {
                        "name": "rating_timing",
                        "in": "query",
                        "description": "rating_timing",
                        "required": true,
                        "schema": {
                            "type": "number",
                            "example": "4"
                        }
                    },
                    {
                        "name": "comments",
                        "in": "query",
                        "description": "comments",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "testing"
                        }
                    },
                    {
                        "name": "title",
                        "in": "query",
                        "description": "title",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "testing"
                        }
                    },
                    {
                        "name": "created_by",
                        "in": "query",
                        "description": "created_by",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "Sam"
                        }
                    },
                    {
                        "name": "user_id",
                        "in": "query",
                        "description": "User Id",
                        "required": true,
                        "schema": {
                            "type": "number",
                            "example": "1"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Add reviews by a customer"
                    },
                    "206": {
                        "description": "validation error"
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "401": {
                        "description": "Unauthorized access"
                    },
                    "404": {
                        "description": "No record found"
                    },
                    "500": {
                        "description": "Internal server error"
                    },
                    "502": {
                        "description": "Bad gateway"
                    },
                    "503": {
                        "description": "Service unavailable"
                    },
                    "504": {
                        "description": "Gateway timeout"
                    }
                },
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            }
        },
...
        "/api/DeleteReview/{id}/{userId}": {
            "delete": {
                "tags": [
                    "Delete Review added by a customer"
                ],
                "summary": "Delete Review",
                "operationId": "988b1386e0023fe6fccee089f6db7882",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "review Id",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    },
                    {
                        "name": "userId",
                        "in": "path",
                        "description": "User Id",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "202": {
                        "description": "Delete Review"
                    },
                    "206": {
                        "description": "User does not match to the review"
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "401": {
                        "description": "Unauthorized access"
                    },
                    "404": {
                        "description": "No record found"
                    },
                    "500": {
                        "description": "Internal server error"
                    },
                    "502": {
                        "description": "Bad gateway"
                    },
                    "503": {
                        "description": "Service unavailable"
                    },
                    "504": {
                        "description": "Gateway timeout"
                    }
                },
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            }
        },
        "/api/ReviewDetail/{id}": {
            "get": {
                "tags": [
                    "Review detail of a customer"
                ],
                "summary": "Review detail of a customer",
                "operationId": "fa5c4e0e0c1b3062e33f0e8fac3f15e2",
                "parameters": [
                    {
                        "name": "review Id",
                        "in": "path",
                        "description": "Review detail of a customer",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Get Review of a Customer"
                    },
                    "206": {
                        "description": "validation error"
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "401": {
                        "description": "Unauthorized access"
                    },
                    "404": {
                        "description": "No record found"
                    },
                    "500": {
                        "description": "Internal server error"
                    },
                    "502": {
                        "description": "Bad gateway"
                    },
                    "503": {
                        "description": "Service unavailable"
                    },
                    "504": {
                        "description": "Gateway timeout"
                    }
                },
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "apiAuth": {
                "type": "http",
                "name": "Token based",
                "in": "header",
                "bearerFormat": "AUTH0",
                "scheme": "bearer"
            }
        }
    },
    "security": [
        []
    ]
}